services/api/requirements.txt
fastapi
prometheus-fastapi-instrumentator
uvicorn
psycopg2-binary
redis
sqlalchemy
python-json-logger
aiofiles
